pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sitegen"
version = "1.0.0"
description = "Gerador de sites SEO para negócios locais brasileiros"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "websockets>=12.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "openai>=1.0.0",
    "google-genai>=0.1.0",
    "requests>=2.31.0",
    "tqdm>=4.66.0",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.9.0",
]

[tool.setuptools.packages.find]
include = ["core", "core.*"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
source = ["core"]
omit = [
    "tests/*",
    "core/imagen_client.py",
    "core/openrouter_client.py",
    "core/template_injector.py",
]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]
